move rewritten broker out of r5 into backend
it's now running within the backend process so it should really be in that project
