+++ src/main/java/com/conveyal/r5/analyst/cluster/AnalysisTask.java
@@ -3,15 +3,13 @@ package com.conveyal.r5.analyst.cluster;
 import com.conveyal.r5.analyst.GridCache;
 import com.conveyal.r5.analyst.PointSet;
 import com.conveyal.r5.analyst.WebMercatorGridPointSetCache;
-import com.conveyal.r5.analyst.broker.WorkerCategory;
+import com.conveyal.r5.analyst.WorkerCategory;
 import com.conveyal.r5.profile.PerTargetPropagater;
 import com.conveyal.r5.profile.ProfileRequest;
 import com.conveyal.r5.transit.TransportNetwork;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
-
-import java.io.OutputStream;
 
 /**
  * Describe an analysis task to be performed, which could be a single point interactive task for a surface of travel times,
